requirements.txt
# ===========================================
# OPTIMIZED REQUIREMENTS.TXT FOR DOCKER
# ===========================================
# This file contains only the packages actually used in the codebase
# Organized by functionality for better understanding

# ===========================================
# CORE WEB FRAMEWORK
# ===========================================
fastapi==0.115.7
uvicorn[standard]==0.34.0
# Note: starlette is included with fastapi, no need to specify separately

# ===========================================
# DATABASE & ORM
# ===========================================
sqlalchemy==2.0.38
alembic==1.14.0

# ===========================================
# AUTHENTICATION & SECURITY
# ===========================================
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
# Note: bcrypt is included with passlib[bcrypt]
cryptography>=3.4.8

# ===========================================
# HTTP & API COMMUNICATION
# ===========================================
python-multipart==0.0.20
requests==2.32.3
httpx==0.27.2

# ===========================================
# TEMPLATE ENGINE
# ===========================================
jinja2==3.1.2
# Note: markupsafe is included with jinja2

# ===========================================
# DATE & TIME UTILITIES
# ===========================================
python-dateutil==2.9.0.post0

# ===========================================
# FILE PROCESSING (ACTUALLY USED)
# ===========================================
python-docx==1.2.0
PyPDF2==3.0.1
# Note: Removed unused PDF libraries (PyMuPDF, pdfplumber, pdfminer.six, docx2txt)

# ===========================================
# AI INTEGRATION
# ===========================================
ollama==0.5.3

# ===========================================
# CONFIGURATION & ENVIRONMENT
# ===========================================
python-dotenv==1.1.0

# ===========================================
# DATA VALIDATION & MODELS
# ===========================================
pydantic==2.10.6
pydantic-settings==2.6.1
typing-extensions==4.13.2

# ===========================================
# EMAIL FUNCTIONALITY
# ===========================================
secure-smtplib==0.1.1

# ===========================================
# OAUTH INTEGRATION
# ===========================================
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.146.0

# ===========================================
# LOGGING & MONITORING
# ===========================================
python-json-logger==2.0.7

# ===========================================
# PRODUCTION SERVER
# ===========================================
gunicorn==23.0.0

# ===========================================
# DEVELOPMENT & TESTING (OPTIONAL)
# ===========================================
# Uncomment these if you need testing in production
# pytest==8.3.5
# pytest-docker==3.1.2

# ===========================================
# REMOVED PACKAGES (NOT USED IN CODEBASE)
# ===========================================
# - pillow (image processing) - not used
# - opencv-python-headless (computer vision) - not used  
# - numpy (numerical computing) - not used
# - pandas (data analysis) - not used
# - PyMuPDF (PDF processing) - not used
# - pdfplumber (PDF processing) - not used
# - pdfminer.six (PDF processing) - not used
# - docx2txt (DOCX processing) - not used